pyproject.toml
[tool.poetry]
name = "repl-nix-smartchessmaster"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flask = "^3.0.3"
flask-socketio = "^5.3.7"
chess = "^1.10.0"
openai = "^1.46.0"
pillow = "^10.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
